Updated unit test case
