+++ src/routers/handlers/check_your_answers/check.your.answers.ts
@@ -53,6 +53,8 @@ export class CheckYourAnswersHandler extends GenericHandler {
         const transactionService = new TransactionService(req.session);
         const paymentUrl: string | undefined = await transactionService.closeTransaction();
 
+        logger.info('Susilaa     ---------------' + paymentUrl);
+
         if (!paymentUrl) {
             logger.debug(`No payment url ${paymentUrl} from closeTransaction, journey redirected to confirmation page`);
             return PrefixedUrls.CONFIRMATION;
