+++ test/routers/check.your.answers.unit.ts
@@ -1,3 +1,5 @@
+jest.mock("../../src/services/external/payment.service");
+
 import { mockSession, resetMockSession } from "../mocks/session.middleware.mock";
 import { mockTransactionService } from "../mocks/transaction.service.mock";
 
@@ -9,14 +11,29 @@ import { AccountValidatorResponse } from "private-api-sdk-node/dist/services/acc
 import { ContextKeys } from "../../src/utils/constants/context.keys";
 import { Account } from "../../src/utils/constants/paymentTypes";
 import { getSessionRequest } from "../mocks/session.mock";
+import { startPaymentsSession } from "../../src/services/external/payment.service";
+import { mockPayment, PAYMENT_JOURNEY_URL } from "../mocks/payment.mock";
+import { Payment } from "@companieshouse/api-sdk-node/dist/services/payment";
+import { ApiResponse } from "@companieshouse/api-sdk-node/dist/services/resource";
+import { logger } from "../../src/utils/logger";
+const mockStartPaymentsSession = startPaymentsSession as jest.Mock;
 
+const mockPaymentHeaders = {
+    header1: "45435435"
+};
+const mockPaymentResponse = {
+    headers: mockPaymentHeaders,
+    httpStatusCode: 200,
+    resource: mockPayment
+} as ApiResponse<Payment>;
+const PAYMENT_URL = "/payment/1234";
 
 describe("Check your answers test", () => {
     beforeEach(() => {
         resetMockSession();
         getSessionRequest();
+        jest.clearAllMocks();
     });
-
     afterEach(() => {
 
     });
@@ -98,5 +115,20 @@ describe("Check your answers test", () => {
         // It Should redirect to the confirmation page
         expect(resp.status).toBe(302);
         expect(resp.headers.location).toBe(PrefixedUrls.CONFIRMATION);
+    });
+
+    it("Should close the transaction and navigate to the payment jouney when receiving a payment url from close transaction", async () => {
+        logger.info("Adhvikkkkkkkkk");
+        mockTransactionService.closeTransaction.mockResolvedValue(PAYMENT_URL);
+        mockStartPaymentsSession.mockResolvedValueOnce(mockPaymentResponse);
+        mockSession.data['signin_info'] = { company_number: "00000000" };
+        mockSession.setExtraData(ContextKeys.COMPANY_NUMBER, "00000000");
+        mockSession.setExtraData(ContextKeys.ACCOUNTS_FILING_ID, "mockAccFilingId");
+        mockSession.setExtraData(ContextKeys.TRANSACTION_ID, "mockTransId");
+        mockSession.data['signin_info']['signed_in'] = 1;
+        const resp = await request(app).post(PrefixedUrls.CHECK_YOUR_ANSWERS);
+        // It Should redirect to the payment journey page
+        expect(resp.status).toBe(302);
+        expect(resp.header.location).toBe(PAYMENT_JOURNEY_URL);
     });
 });
